day5 (had to look at other solutions but yeah lol)
My initial solution was way too slow so I had to look up
other people's solutions to find out going backwards is
faster. Then I had to use someone else's code to find out
my solution, which revealed that it was really low and my
code was passing by it. Then the fix was clear and I got the
solution with my code (very quickly because the solution was
literally the minimum result lol)